fix: resolve type checker errors and security scan findings (Refs PMAT-112)
- Add explicit type casts for dict value access across 8 modules
- Fix MD5 usage with usedforsecurity=False in video.py
- Add nosec annotations for intentional pickle.loads in reproducibility.py
- Add nosec for hardcoded bind address and tmp path

All quality gates pass: lint, typecheck, 16372 tests, 98% coverage, security.

+++ src/hf_gtc/rag/chunking.py
@@ -927,8 +927,8 @@ def get_recommended_chunking_config(
     }
 
     adj = type_adjustments[document_type]
-    chunk_size = int(base["chunk_size"] * adj["size_mult"])
-    overlap_size = int(base["overlap_size"] * adj["size_mult"])
+    chunk_size = int(int(base["chunk_size"]) * float(adj["size_mult"]))
+    overlap_size = int(int(base["overlap_size"]) * float(adj["size_mult"]))
     strategy = adj["strategy"]
 
     return ChunkConfig(
